backend/src/sphinx-bot/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
cartesia==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.5.0
daily-python==0.16.1
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
hume==0.7.13
idna==3.10
iterators==0.2.0
jiter==0.9.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.7
mpmath==1.3.0
multidict==6.3.2
numba==0.61.0
numpy==1.26.4
onnxruntime==1.21.0
openai==1.67.0
packaging==24.2
pillow==11.1.0
Pint==0.24.4
pipecat-ai==0.0.62
pipecat-ai-flows==0.0.16
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
resampy==0.4.3
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.46.1
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
yarl==1.19.0
